refactor(e2e): extract shared utilities from runtime-*.ts
Move reusable utilities out of runtime-deploy.ts and
runtime-config.ts into shared modules:

helper.ts:
- run() — shell command execution with stdout/stderr capture
- discoverRouterBase() — OpenShift cluster router base discovery
- ImageRef interface, imageRefToString(), parseCatalogIndexImage()
  — image reference parsing utilities

kube-client.ts (KubeClient class):
- createNamespace() — 409-safe namespace creation
- deleteNamespaceIfExists() — 404-safe deletion wrapping
  deleteNamespaceAndWait()

runtime-config.ts re-exports ImageRef, imageRefToString, and
parseCatalogIndexImage from helper.ts so existing callers are
unaffected.

runtime-deploy.ts now imports all shared utilities instead of
defining them locally, reducing file size by ~100 lines.
